Move client and log package
