+++ package.json
@@ -3,11 +3,13 @@
   "license": "Apache-2.0",
   "author": "Kinvey, Inc.",
   "scripts": {
-    "test": "mocha --require babel-register src/**/*.spec.js"
+    "test": "mocha --require babel-register 'src/**/*.spec.js'"
   },
   "dependencies": {
     "es6-promise": "~4.1.1",
+    "fast-memory-cache": "~2.0.4",
     "lodash": "~4.17.4",
+    "loglevel": "~1.5.1",
     "qs": "~6.5.1",
     "rxjs": "~5.5.2"
   },
